requirements.txt
pandas
numpy
scipy
statsmodels
scikit-learn
matplotlib
seaborn
pyarrow
pyyaml
openpyxl
joblib
tqdm
xgboost
lightgbm
catboost
